requirements.txt
haystack-ai
mistral-haystack
jupyter
dotenv
trafilatura